chang the octal number prototybe

+++ main.h
@@ -16,7 +16,7 @@ int print_unsigned(va_list args, int flags, int length);
 int print_octal(va_list args, int flags, int length);
 int print_hex(va_list args, int uppercase, int flags, int length);
 int print_unsigned_number(unsigned long n);
-int print_octal_number(unsigned int n);
+int print_octal_number(unsigned long n);
 int print_hex_number(unsigned int n, int uppercase);
 int print_string_escaped(va_list args);
 int print_pointer(va_list args);
